pqcrypto_mldsa

Module mldsa87

ml-dsa-87

These bindings use the clean version from PQClean

Example

// if using pqcrypto-mldsa
use pqcrypto_mldsa::mldsa87::*;
// or if using the pqcrypto crate:
// use pqcrypto::sign::mldsa87::*;
let message = vec![0, 1, 2, 3, 4, 5];
let (pk, sk) = keypair();
let sm = sign(&message, &sk);
let verifiedmsg = open(&sm, &pk).unwrap();
assert!(verifiedmsg == message);

Structs

DetachedSignature

PublicKey

SecretKey

SignedMessage

Functions

detached_sign

Create a detached signature on the message

detached_sign_ctx

Create a detached signature on the message

keypair

Generate a ml-dsa-87 keypair

open

Open the signed message and if verification succeeds return the message

open_ctx

Open the signed message and if verification succeeds return the message

public_key_bytes

Get the number of bytes for a public key

secret_key_bytes

Get the number of bytes for a secret key

sign

Sign the message and return the signed message.

sign_ctx

Sign the message for the provided context and return the signed message.

signature_bytes

Get the number of bytes that a signature occupies

verify_detached_signature

Verify the detached signature

verify_detached_signature_ctx

Verify the detached signature